+++ src/microapi/functions_include.php
@@ -1,0 +1,12 @@
+<?php
+/**
+ * Created by IntelliJ IDEA.
+ * User: alex
+ * Date: 02.08.17
+ * Time: 14:12
+ */
+
+// Don't redefine the functions if included multiple times.
+if (!function_exists('\microapi\send_response')) {
+    require __DIR__ . '/functions.php';
+}