+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+#0.0.26-alpha
+- added DefaultResponseFactory. Should minimize calling of the clone() in \Psr\Http\Message\ResponseInterface::with* methods
+- added examples for microapi-framework usage in tests/functional/router.php 
+- added sending Response to the client
+- event handlers now can be inserted before or after previous
+
 #0.0.23-alpha
 - PSR-7 integration completed (but still need some improvements)